requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
sentence-transformers
pydantic
chromadb
python-dotenv
tqdm